[TM-1717] add impact story policy and move to correct folder
